validacion conexion redis

+++ src/services/redis.js
@@ -1,60 +1,94 @@
 'use strict'
 
-module.exports = function(URL_CONECTION) {
+module.exports = function (URL_CONECTION) {
   const Redis = require("async-redis")
-  const redis = Redis.createClient(URL_CONECTION)
-  
+  const redis = Redis.createClient(URL_CONECTION).on('error', function (err) {
+    console.log(URL_CONECTION + " " + err);
+  }).on('connect', function () {
+    console.log('Redis connected ' + URL_CONECTION);
+  })
   let service = {}
-  
+
   service.set = function (clave, value) {
-    return redis.set(clave, value)
+    if (redis.connected)
+      return redis.set(clave, value)
+    else
+      return false
   }
 
   service.setWithExpire = function (clave, value, expire) {
-    if(expire)
-      return redis.set(clave,value,'EX', expire)
-    else
-      return redis.set(clave,value)
+    if (redis.connected) {
+      if (expire)
+        return redis.set(clave, value, 'EX', expire)
+      else
+        return redis.set(clave, value)
+    } else
+      return false
   }
-  
+
   service.get = async function (clave) {
-    let resultado = await redis.get(clave)
-    return resultado
+    if (redis.connected) {
+      let resultado = await redis.get(clave)
+      return resultado
+    } else
+      return false
   }
 
   service.setJson = function (clave, value) {
-    return redis.set(clave, JSON.stringify(value))
+    if (redis.connected)
+      return redis.set(clave, JSON.stringify(value))
+    else
+      return false
   }
-  
+
   service.getJson = async function (clave) {
-    let resultado = await redis.get(clave)
-    return JSON.parse(resultado)
+    if (redis.connected) {
+      let resultado = await redis.get(clave)
+      return JSON.parse(resultado)
+    } else
+      return false
   }
 
   service.setObject = function (clave, value) {
-    return redis.hmset(clave, value)
+    if (redis.connected) {
+      return redis.hmset(clave, value)
+    }
+    return false
   }
-  
+
   service.getObject = async function (clave) {
-    let resultado = await redis.hgetall(clave)
-    return resultado
+    if (redis.connected) {
+      let resultado = await redis.hgetall(clave)
+      return resultado
+    } else
+      return false
   }
 
   service.deleteByPattern = async function (pattern) {
-    const  keys = await redis.keys (pattern)
-    keys.forEach(key => {
-      redis.del(key)
-    })
-    return keys
+    if (redis.connected) {
+      const keys = await redis.keys(pattern)
+      keys.forEach(key => {
+        redis.del(key)
+      })
+      return keys
+    } else
+      return false
   }
 
   service.deleteByKey = async function (key) {
-    return await redis.del(key)
+    if (redis.connected)
+      return await redis.del(key)
+    else
+      return false
   }
 
   // Clear All Keys
   service.clearAll = async function () {
-    return await redis.flushall();
+    if (redis.connected) {
+      return await redis.flushall();
+    }
+    else
+      return false
   }
 
   return service
